+++ app/src/organisms/ProtocolSetup/RunSetupCard/ModuleSetup/HeaterShakerSetupWizard/HeaterShakerBanner.tsx
@@ -24,15 +24,19 @@ export function HeaterShakerBanner(
     <Banner title={t('attach_heater_shaker_to_deck', { name: displayName })}>
       {modules.map((module, index) => (
         <React.Fragment key={index}>
-          {wizardId === module.moduleId &&
-            module.attachedModuleMatch?.moduleType ===
-              HEATERSHAKER_MODULE_TYPE && (
-              <HeaterShakerWizard
-                onCloseClick={() => setWizardId(null)}
-                moduleFromProtocol={module}
-                attachedModule={module.attachedModuleMatch}
-              />
-            )}
+          {wizardId === module.moduleId && (
+            <HeaterShakerWizard
+              onCloseClick={() => setWizardId(null)}
+              moduleFromProtocol={module}
+              attachedModule={
+                module.attachedModuleMatch != null &&
+                module.attachedModuleMatch?.moduleType ===
+                  HEATERSHAKER_MODULE_TYPE
+                  ? module.attachedModuleMatch
+                  : null
+              }
+            />
+          )}
           {index > 0 && <Divider color={COLORS.medGreyEnabled} />}
           <BannerItem
             title={t('module_in_slot', {